Added Browser code editor
NOT SAFE FOR DEPLOYMENT
Added Browser code editor support for Javascript, Python, C++, and Java via Docker. The code is run by putting the inputted code into a temporary file and run in a console depending on the selected version. This has security risks that have yet to be prevented, so do not publish this version under any circumstance. Text can be inputted to the website using a CodeMirror text editor. Moving forward, a Github Pages method is no longer avaliable due to the website being dynamic